# server/app/src/tsoa-build/openapi.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"AuthErrorResponseBody": {
				"$ref": "#/components/schemas/ErrorResponseBody_AuthErrorResponseContent.AuthErrorBodyMessageT_",
				"description": "The response body for an {@link AuthError}"
			},
...
			"ValidationErrorResponseBody": {
				"$ref": "#/components/schemas/ErrorResponseBody_ValidationErrorResponseContent.ValidationErrorMessageT_",
				"description": "Represents the response body for a validation error\n\nUsed by internal {@link QueryValidationError} and {@link ClassValidationError},\nas well as `sequelize`.{@link ValidationError} and `tsoa`.`ValidateError`"
			},
...
			"DataResponseBody_Array_ReadResponseContent_UserModelDTO.password.never__.User.FoundResponseMessageT_": {
				"description": "Represents the response body of a data response",
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/ReadResponseContent_UserModelDTO.password.never_"
						},
						"type": "array"
					},
					"message": {
						"$ref": "#/components/schemas/FoundResponseMessageT"
					},
					"metadata": {
						"$ref": "#/components/schemas/MetaData_User_"
					}
				},
				"required": [
					"data",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserReadManyResponse": {
				"$ref": "#/components/schemas/DataResponseBody_Array_ReadResponseContent_UserModelDTO.password.never__.User.FoundResponseMessageT_",
				"description": "Possible responses for {@link UserController.getAllUsers}"
			},
			"OrderByFieldValues": {
				"description": "Possible values for the filter query paramater on controller routes",
				"enum": [
					"createdAt",
					"updatedAt",
					"-createdAt",
					"-updatedAt"
				],
				"type": "string"
			},
...
			"NotFoundResponseBody": {
				"$ref": "#/components/schemas/ErrorResponseBody_NotFoundResponseContent.NotFoundMessageT_",
				"description": "The type of {@link ErrorResponseBody} for not found responses"
			},
			"UserNotFoundResponseBody": {
				"$ref": "#/components/schemas/NotFoundResponseBody",
				"description": "Response type for this controllers object not found response"
			},
...
			"UserReadResponse": {
				"$ref": "#/components/schemas/DataResponseBody_ReadResponseContent_UserModelDTO.password.never_.User.FoundResponseMessageT_",
				"description": "Possible responses for {@link UserController.getUser}"
			},
			"AppModelIdT": {
				"type": "number",
				"format": "double",
				"description": "This is the type of {@link AppModel.id} - used to ensure its correct when used in routers etc."
			},
...
			"UserCreateResponse": {
				"$ref": "#/components/schemas/DataResponseBody_CreateResponseContent_UserModelDTO.password.never_.User.CreatedResponseMessageT_",
				"description": "Possible responses for {@link UserController.createUser}"
			},
...
			"CreateRequestBody_UserModelDTO.isAdmin-or-isStaff-or-fullName.profileImageURL_": {
				"$ref": "#/components/schemas/TransformKeys_UserModelDTO.Extract_isAdmin-or-isStaff-or-fullName-or-CreateRequestBodyOmittedKeysT.keyofUserModelDTO_.profileImageURL.never_"
			},
			"UserCreateRequest": {
				"$ref": "#/components/schemas/CreateRequestBody_UserModelDTO.isAdmin-or-isStaff-or-fullName.profileImageURL_",
				"description": "Request body format for {@link UserController.createUser}'s `user` body parameter"
			},
...
			"UserUpdateResponse": {
				"$ref": "#/components/schemas/DataResponseBody_UpdateResponseContent_UserModelDTO.password.never_.User.UpdatedResponseMessageT_",
				"description": "Possible responses for {@link UserController.updateUser}"
			},
...
			"UserUpdateRequest": {
				"$ref": "#/components/schemas/UpdateRequestBody_UserModelDTO.isAdmin-or-isStaff-or-fullName.profileText_",
				"description": "Request body format for {@link UserController.updateUser}'s `user` body parameter"
			},
...
			"UserDeleteResponse": {
				"$ref": "#/components/schemas/DeleteResponseBody_User_",
				"description": "Possible responses for {@link UserController.deleteUser}"
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "api_key",
				"in": "query"
			},
			"oauth2": {
				"type": "oauth2",
				"flows": {
					"password": {
						"tokenUrl": "/auth/login",
						"scopes": {
							"admin": "modify things",
							"staff": "modify some things",
							"user": "read things"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": ""
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserReadManyResponse"
								}
							}
						}
					},
					"401": {
						"description": "Not Authenticated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"403": {
						"description": "Not Authorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserNotFoundResponseBody"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponseBody"
								}
							}
						}
					}
				},
				"description": "Retrieves a list of all users, matching the query parameters if provided.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"description": "- Retrieve records starting from this offset",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- Retrieve up to this many records",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- Order records by this attribute",
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderByFieldValues"
						}
					},
					{
						"description": "- Stringified JSON where the keys represent a property of a user and values represent a text based filter for the property",
						"in": "query",
						"name": "filters",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserCreateResponse"
								}
							}
						}
					},
					"401": {
						"description": "Not Authenticated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"403": {
						"description": "Not Authorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponseBody"
								}
							}
						}
					}
				},
				"description": "Creates a new user",
				"tags": [
					"User"
				],
				"security": [
					{
						"api_key": []
					},
					{
						"oauth2": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "- The posted body object, containing the creation attributes of UserModel",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreateRequest",
								"description": "- The posted body object, containing the creation attributes of UserModel"
							}
						}
					}
				}
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserReadResponse"
								}
							}
						}
					},
					"401": {
						"description": "Not Authenticated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"403": {
						"description": "Not Authorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserNotFoundResponseBody"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponseBody"
								}
							}
						}
					}
				},
				"description": "Retrieves the details of an existing user with a matching id",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"description": "- The identifier of the user to get",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/AppModelIdT"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserUpdateResponse"
								}
							}
						}
					},
					"401": {
						"description": "Not Authenticated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"403": {
						"description": "Not Authorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserNotFoundResponseBody"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponseBody"
								}
							}
						}
					}
				},
				"description": "Update a user with a matching id",
				"tags": [
					"User"
				],
				"security": [
					{
						"api_key": []
					},
					{
						"oauth2": [
							"user"
						]
					}
				],
				"parameters": [
					{
						"description": "- The identifier of the user to update",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/AppModelIdT"
						}
					}
				],
				"requestBody": {
					"description": "- The posted body object, containing some or all modifiable attributes of UserModel",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdateRequest",
								"description": "- The posted body object, containing some or all modifiable attributes of UserModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDeleteResponse"
								}
							}
						}
					},
					"401": {
						"description": "Not Authenticated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"403": {
						"description": "Not Authorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserNotFoundResponseBody"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponseBody"
								}
							}
						}
					}
				},
				"description": "Delete the user with the matching id",
				"tags": [
					"User"
				],
				"security": [
					{
						"api_key": []
					},
					{
						"oauth2": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- The identifier of the user to delete",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/AppModelIdT"
						}
					},
					{
						"description": "- If true the user will actually be deleted from the database rather than a soft delete (handled by sequelize)",
						"in": "query",
						"name": "force",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
	"basePath": "/",
...
	"securityDefinitions": {
		"api_key": {
			"type": "apiKey",
			"name": "api_key",
			"in": "query"
		},
		"oauth2": {
			"type": "oauth2",
			"tokenUrl": "/auth/login",
			"flow": "password",
			"scopes": {
				"admin": "modify things",
				"staff": "modify some things",
				"user": "read things"
			}
		}
	}
}